+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No changes yet.
+### Added
+- Generated exceptions now include an `ErrorName()` method that returns the
+  name of the exception as defined in the Thrift file.
 
 ## [1.20.2] - 2019-10-17
 ### Fixed